refactor(checker): split concrete indexed-access diagnostics
